move includes into dicom/ subdir
So we can now always include our header with
`<dicom/dicom.h>` in all source files.
