+++ include/dicom/version.h.in
